+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Request location on mount
+- Remove filter button
 
 ## [0.5.0] - 2025-02-23